requirements.txt
Flask
requests
Flask-Login
python-dotenv
Werkzeug